+++ CHANGELOG.md
@@ -4,4 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2020-06-16
+### Changes
+- Adds ability to set more config values like cookie domain, cookies paths etc.
+- Adds ability to connect to https instance
+- Changes how SuperTokens hosts are given